+++ packages/create-lila-app/bin/create-lila-app.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-
-require('../lib'); // eslint-disable-line import/no-unresolved