+++ packages/lila-rollup-config/example/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  rules: {
-    'no-console': 0,
-    'import/extensions': 0,
-  },
-};