+++ packages/lila-rollup-config/example/app/pkgs/second/src/index.js
@@ -1,3 +1,0 @@
-console.log('second');
-
-export default {};