+++ packages/lila-rollup-config/example/app/src/flow.js
@@ -1,5 +1,0 @@
-// @flow
-
-export default (num: number) => {
-  console.log(num + num);
-};