+++ packages/lila-webpack-config/src/data.js
@@ -1,4 +1,0 @@
-export const baseType = 'base';
-export const reactType = 'react';
-export const vueType = 'vue';
-export const reactVueType = 'react+vue';