+++ packages/lila-webpack-lib-config/example/app/demo-vue/Jsx.vue
@@ -1,9 +1,0 @@
-<script>
-export default {
-  name: 'Jsx',
-  functional: true,
-  render() {
-    return <h1>Jsx</h1>;
-  },
-};
-</script>